+++ src/warc2zim/main.py
@@ -19,8 +19,10 @@ If the WARC contains multiple entries for the same URL, only the first entry is 
 """
 
 import os
+import sys
 import pathlib
 import logging
+import tempfile
 import mimetypes
 import datetime
 import re
@@ -248,15 +250,15 @@ class WARC2Zim:
 
         self.full_filename = os.path.join(self.output, self.zim_file)
 
+        # ensure output file is writable
+        with tempfile.NamedTemporaryFile(dir=self.output, delete=True) as fh:
+            logger.debug(f"Confirming output is writable using {fh.name}")
+
         self.inputs = args.inputs
         self.replay_viewer_source = args.replay_viewer_source
 
         self.main_url = args.url
-        self.include_all = args.include_all
         self.include_domains = args.include_domains
-        if self.main_url:
-            if not self.include_all and not self.include_domains:
-                self.include_domains = [urlsplit(self.main_url).netloc]
 
         self.favicon_url = args.favicon
         self.language = args.lang
@@ -305,6 +307,12 @@ class WARC2Zim:
         return env
 
     def run(self):
+        if not self.inputs:
+            logger.info(
+                "Arguments valid, no inputs to process. Exiting with error code 100"
+            )
+            return 100
+
         self.find_main_page_metadata()
 
         # make sure Language metadata is ISO-639-3 and setup translations
@@ -377,8 +385,6 @@ class WARC2Zim:
                 )
             ):
                 self.main_url = url
-                if not self.include_all and not self.include_domains:
-                    self.include_domains = [urlsplit(self.main_url).netloc]
 
             if self.main_url != url:
                 continue
@@ -489,8 +495,8 @@ class WARC2Zim:
             logger.debug("Skipping duplicate {0}, already added to ZIM".format(url))
             return
 
-        # if not include_all, only include urls from main_url domain or subdomain
-        if not self.include_all:
+        # if include_domains is set, only include urls from those domains
+        if self.include_domains:
             parts = urlsplit(url)
             if not any(
                 parts.netloc.endswith(domain) for domain in self.include_domains
@@ -549,7 +555,7 @@ def warc2zim(args=None):
 
     parser.add_argument(
         "inputs",
-        nargs="+",
+        nargs="*",
         help="""Paths of directories and/or files to be included in
                                 the WARC file.""",
     )
@@ -567,17 +573,10 @@ def warc2zim(args=None):
     )
 
     parser.add_argument(
-        "-a",
-        "--include-all",
-        action="store_true",
-        help="If set, include all URLs in ZIM, not just those specified in --include-domains",
-    )
-
-    parser.add_argument(
         "-i",
         "--include-domains",
         action="append",
-        help="List of domains that should be included. Not used if --include-all is set. Defaults to domain of the main url",
+        help="Limit ZIM file to URLs from only certain domains. If not set, all URLs in the input WARCs are included.",
     )
 
     parser.add_argument(
@@ -629,4 +628,4 @@ def get_version():
 
 # ============================================================================
 if __name__ == "__main__":  # pragma: no cover
-    warc2zim()
+    sys.exit(warc2zim())